Data-Cleaning Automation
Description:

The CSV Data Cleaner is a Python-based application that automates the process of cleaning missing data in CSV files. Users can fill missing values in numeric columns with the mean, median, or mode of the respective columns. This tool simplifies the process of data preprocessing, making it faster and more efficient.

Features:

-Fill Empty Cells with Mean: Fills missing values in numeric columns with the mean of that column. -Fill Empty Cells with Median: Fills missing values in numeric columns with the median of that column. -Fill Empty Cells with Mode: Fills missing values in numeric columns with the mode (most frequent value) of that column. -Automatic File Saving: After cleaning, the file is saved with a timestamped filename for easy tracking. -Exit Option: Allows the user to cancel the operation and exit.

Author:
-[Mohammad Mohid]
-[LinkedIn Profile](https://www.linkedin.com/in/mohammad-mohid-162585361?utm_source=share&utm_campaign=share_via&utm_content=profile&utm_medium=android_app)